cli/requirements.txt
redis==5.0.1
click==8.1.7
rich==13.7.0
pydantic==2.11.7
python-dotenv==1.0.0
tabulate==0.9.0
